requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.1.20
langchain-google-genai==1.0.3
sentence-transformers==3.1.1
faiss-cpu==1.12.0  # Available version
pdfminer.six==20240706
numpy==2.1.3
pyyaml==6.0.1
pytest==8.3.3